#213: Create useProfileSections-hook
